requirements.txt
aniso8601==8.0.0
attrs==19.3.0
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Faker==4.1.1
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
future==0.18.2
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-bigquery==1.25.0
google-cloud-core==1.3.0
google-cloud-pubsub==1.6.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.29.0
gunicorn==20.0.4
httpie==2.2.0
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.6.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
protobuf==3.12.2
py==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybigquery==0.4.15
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyotp==2.3.0
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.17
text-unidecode==1.3
urllib3==1.25.9
wcwidth==0.2.4
Werkzeug==1.0.1
zipp==3.1.0
